Move ODataOperationFilter to separated file
