+++ src/Templates/TodoTemplate/Bit.TodoTemplate/Api/Extensions/IServiceCollectionExtensions.cs
@@ -2,7 +2,6 @@
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
@@ -99,106 +98,6 @@ public static class IServiceCollectionExtensions
         });
 
         services.AddAuthorization();
-    }
-
-    /// <summary>
-    /// https://docs.microsoft.com/en-us/odata/concepts/queryoptions-overview
-    /// </summary>
-    public class ODataOperationFilter : IOperationFilter
-    {
-        public void Apply(OpenApiOperation operation, OperationFilterContext context)
-        {
-            if (operation.Parameters == null) operation.Parameters = new List<OpenApiParameter>();
-
-            var descriptor = context.ApiDescription.ActionDescriptor as ControllerActionDescriptor;
-
-            if (descriptor != null && descriptor.FilterDescriptors.Any(filter => filter.Filter is EnableQueryAttribute))
-            {
-                operation.Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "$select",
-                    In = ParameterLocation.Query,
-                    Schema = new OpenApiSchema
-                    {
-                        Type = "string",
-                    },
-                    Description = "Returns only the selected properties. (ex. FirstName, LastName)",
-                    Required = false
-                });
-
-                operation.Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "$expand",
-                    In = ParameterLocation.Query,
-                    Schema = new OpenApiSchema
-                    {
-                        Type = "string",
-                    },
-                    Description = "Include only the selected objects. (ex. Childrens, Locations)",
-                    Required = false
-                });
-
-                operation.Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "$filter",
-                    In = ParameterLocation.Query,
-                    Schema = new OpenApiSchema
-                    {
-                        Type = "string",
-                    },
-                    Description = "Filter the response with OData filter queries.",
-                    Required = false
-                });
-
-                operation.Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "$search",
-                    In = ParameterLocation.Query,
-                    Schema = new OpenApiSchema
-                    {
-                        Type = "string",
-                    },
-                    Description = "Filter the response with OData search queries.",
-                    Required = false
-                });
-
-                operation.Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "$top",
-                    In = ParameterLocation.Query,
-                    Schema = new OpenApiSchema
-                    {
-                        Type = "int",
-                    },
-                    Description = "Number of objects to return. (ex. 25)",
-                    Required = false
-                });
-
-                operation.Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "$skip",
-                    In = ParameterLocation.Query,
-                    Schema = new OpenApiSchema
-                    {
-                        Type = "int",
-                    },
-                    Description = "Number of objects to skip in the current order (ex. 50)",
-                    Required = false
-                });
-
-                operation.Parameters.Add(new OpenApiParameter()
-                {
-                    Name = "$orderby",
-                    In = ParameterLocation.Query,
-                    Schema = new OpenApiSchema
-                    {
-                        Type = "string",
-                    },
-                    Description = "Define the order by one or more fields (ex. LastModified)",
-                    Required = false
-                });
-            }
-        }
     }
 
     public static void AddSwaggerGen(this IServiceCollection services)
